Fixed Entrepreneur API

+++ src/main/java/com/jumpstartup/Database/EntrepreneurDatabase.java
@@ -2,6 +2,8 @@ package com.jumpstartup.Database;
 
 import com.jumpstartup.Connection.DatabaseConnector;
 import com.jumpstartup.Entrepreneur.EntrepreneurBean;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.sql.Connection;
 
@@ -11,7 +13,8 @@ import java.sql.SQLException;
 
 
 public class EntrepreneurDatabase {
-    
+
+    private static final Logger logger = LoggerFactory.getLogger(EntrepreneurDatabase.class);
     public boolean addEntrepreneur(EntrepreneurBean entrepreneur) {
         Connection connection = null;
         try {
@@ -29,7 +32,60 @@ public class EntrepreneurDatabase {
                 return false;
             }
         } catch (SQLException e) {
-            //Logging statement here System.out.println("Error while trying to add entrepreneur: " + e.getMessage());
+            logger.error("Error while trying to add entrepreneur: {}", e.getMessage());
+            return false;
+        } finally {
+            DatabaseConnector.closeConnection(connection);
+        }
+    }
+
+    public boolean addEducation(String uuid, String institution, String degree, String major, int year_of_completion) {
+        Connection connection = null;
+        try {
+            connection = DatabaseConnector.getConnection();
+            String sql = "INSERT INTO Education (UUID, institution, degree, major, year_of_completion) VALUES (?, ?, ?, ?, ?)";
+            PreparedStatement statement = connection.prepareStatement(sql);
+            statement.setString(1, uuid);
+            statement.setString(2, institution);
+            statement.setString(3, degree);
+            statement.setString(4, major);
+            statement.setInt(5, year_of_completion);
+
+            int rowsAffected = statement.executeUpdate();
+            if (rowsAffected > 0) {
+                logger.info("Education added successfully for entrepreneur with UUID: {}", uuid);
+                return true;
+            } else {
+                logger.warn("Failed to add education for entrepreneur with UUID: {}", uuid);
+                return false;
+            }
+        } catch (SQLException e) {
+            logger.error("Error while trying to add education: {}", e.getMessage());
+            return false;
+        } finally {
+            DatabaseConnector.closeConnection(connection);
+        }
+    }
+
+    public boolean addWorkExperience(String uuid, String work_experience) {
+        Connection connection = null;
+        try {
+            connection = DatabaseConnector.getConnection();
+            String sql = "INSERT INTO Work_Experience (UUID, work_experience) VALUES (?, ?)";
+            PreparedStatement statement = connection.prepareStatement(sql);
+            statement.setString(1, uuid);
+            statement.setString(2, work_experience);
+
+            int rowsAffected = statement.executeUpdate();
+            if (rowsAffected > 0) {
+                logger.info("Work experience added successfully for entrepreneur with UUID: {}", uuid);
+                return true;
+            } else {
+                logger.warn("Failed to add work experience for entrepreneur with UUID: {}", uuid);
+                return false;
+            }
+        } catch (SQLException e) {
+            logger.error("Error while trying to add work experience: {}", e.getMessage());
             return false;
         } finally {
             DatabaseConnector.closeConnection(connection);
@@ -46,18 +102,39 @@ public class EntrepreneurDatabase {
             PreparedStatement statement = connection.prepareStatement(sql);
             statement.setString(1, entrepreneur.getPhone_number());
             statement.setString(2, entrepreneur.getDomain());
-            statement.setString(5, uuid);
+            statement.setString(3, uuid);
 
             int rowsAffected = statement.executeUpdate();
 
+            // update education information
+            String sqlEducation = "UPDATE Education SET institution = ?, degree = ?, major = ?, year_of_completion = ? WHERE uuid = ?";
+            PreparedStatement statementEducation = connection.prepareStatement(sqlEducation);
+            statementEducation.setString(1, entrepreneur.getInstitution());
+            statementEducation.setString(2, entrepreneur.getDegree());
+            statementEducation.setString(3, entrepreneur.getMajor());
+            statementEducation.setInt(4, entrepreneur.getYear_of_completion());
+            statementEducation.setString(5, uuid);
+
+            int rowsAffectedEducation = statementEducation.executeUpdate();
+
+            // update work experience information
+            String sqlWorkExperience = "UPDATE Work_Experience SET work_experience = ? WHERE uuid = ?";
+            PreparedStatement statementWorkExperience = connection.prepareStatement(sqlWorkExperience);
+            statementWorkExperience.setString(1, entrepreneur.getWork_experience());
+            statementWorkExperience.setString(2, uuid);
+
+            int rowsAffectedWorkExperience = statementWorkExperience.executeUpdate();
+
             // check if updates were successful
-            if (rowsAffected > 0 ) {
+            if (rowsAffected > 0 && rowsAffectedEducation > 0 && rowsAffectedWorkExperience > 0) {
+                logger.info("Entrepreneur with UUID {} has been updated successfully.", uuid);
                 return true;
             } else {
+                logger.warn("Failed to update entrepreneur with UUID {}.", uuid);
                 return false;
             }
         } catch (SQLException e) {
-            //Logging statement here System.out.println("Error while trying to update entrepreneur: " + e.getMessage());
+            logger.error("Error while trying to update entrepreneur: {}", e.getMessage());
             return false;
         } finally {
             DatabaseConnector.closeConnection(connection);
@@ -69,18 +146,38 @@ public class EntrepreneurDatabase {
         try {
             connection = DatabaseConnector.getConnection();
 
+            String educationSql = "DELETE FROM Education WHERE uuid = ?";
+            PreparedStatement educationStatement = connection.prepareStatement(educationSql);
+            educationStatement.setString(1, UUID);
+            int educationRowsAffected = educationStatement.executeUpdate();
+            if (educationRowsAffected <= 0) {
+                logger.warn("No education rows affected while deleting entrepreneur with UUID: {}", UUID);
+                return false;
+            }
+            
+            String workExperienceSql = "DELETE FROM Work_Experience WHERE uuid = ?";
+            PreparedStatement workExperienceStatement = connection.prepareStatement(workExperienceSql);
+            workExperienceStatement.setString(1, UUID);
+            int workExperienceRowsAffected = workExperienceStatement.executeUpdate();
+            if (workExperienceRowsAffected <= 0) {
+                logger.warn("No work experience rows affected while deleting entrepreneur with UUID: {}", UUID);
+                return false;
+            }
+            
             // Delete entrepreneur
             String sql = "DELETE FROM Entrepreneur WHERE uuid = ?";
             PreparedStatement statement = connection.prepareStatement(sql);
             statement.setString(1, UUID);
             int rowsAffected = statement.executeUpdate();
             if (rowsAffected > 0 ) {
+                logger.info("Entrepreneur with UUID {} deleted successfully", UUID);
                 return true;
             } else {
+                logger.warn("No table 'Entrepreneur' rows affected while deleting entrepreneur with UUID: {}", UUID);
                 return false;
             }
         } catch (SQLException e) {
-            //Logging statement here System.out.println("Error while trying to delete entrepreneur: " + e.getMessage());
+            logger.error("Error while trying to delete entrepreneur with UUID: {}, error message: {}", UUID, e.getMessage());
             return false;
         } finally {
             DatabaseConnector.closeConnection(connection);
@@ -102,6 +199,22 @@ public class EntrepreneurDatabase {
                 entrepreneur.setPhone_number(result.getString("phone_number"));
                 entrepreneur.setDomain(result.getString("domain"));
 
+                statement = connection.prepareStatement("SELECT * FROM education WHERE uuid = ?");
+                statement.setString(1, UUID);
+                result = statement.executeQuery();
+                if (result.next()) {
+                    entrepreneur.setInstitution(result.getString("institution"));
+                    entrepreneur.setDegree(result.getString("degree"));
+                    entrepreneur.setMajor(result.getString("major"));
+                    entrepreneur.setYear_of_completion(result.getInt("year_of_completion"));
+                }
+
+                statement = connection.prepareStatement("SELECT * FROM work_experience WHERE uuid = ?");
+                statement.setString(1, UUID);
+                result = statement.executeQuery();
+                if (result.next()) {
+                    entrepreneur.setWork_experience(result.getString("work_experience"));
+                }
             }
 
         } catch (SQLException e) {
